Materialize upgrade to Spring Boot 3

+++ materialize/src/main/java/com/giraone/jobs/materialize/persistence/JobRecordRepository.java
@@ -17,7 +17,7 @@ public interface JobRecordRepository extends R2dbcRepository<JobRecord, String> 
         INSERT INTO job_record
         VALUES(:id, :jobAcceptedTimestamp, :lastEventTimestamp, :lastRecordUpdateTimestamp, :status, :pausedBucketKey, :processId)
         """)
-    Mono<Integer> insert(
+    Mono<Long> insert(
         long id, Instant jobAcceptedTimestamp, Instant lastEventTimestamp,
         Instant lastRecordUpdateTimestamp, String status, String pausedBucketKey, long processId);
 
@@ -27,7 +27,7 @@ public interface JobRecordRepository extends R2dbcRepository<JobRecord, String> 
         VALUES(:id, :jobAcceptedTimestamp, :lastEventTimestamp, :lastRecordUpdateTimestamp, :status, :pausedBucketKey, :processId)
         ON CONFLICT (id) DO NOTHING
         """)
-    Mono<Integer> insertIgnoreConflict(
+    Mono<Long> insertIgnoreConflict(
         long id, Instant jobAcceptedTimestamp, Instant lastEventTimestamp,
         Instant lastRecordUpdateTimestamp, String status, String pausedBucketKey, long processId);
 
@@ -38,10 +38,12 @@ public interface JobRecordRepository extends R2dbcRepository<JobRecord, String> 
         ON CONFLICT (id) DO UPDATE SET
         last_event_timestamp=:lastEventTimestamp, last_record_update_timestamp=:lastRecordUpdateTimestamp, status=:status, paused_bucket_key=:pausedBucketKey
         """)
-    Mono<Integer> insertOnConflictUpdate(
+    Mono<Long> insertOnConflictUpdate(
         long id, Instant jobAcceptedTimestamp, Instant lastEventTimestamp,
         Instant lastRecordUpdateTimestamp, String status, String pausedBucketKey, long processId);
 
+    /*
+    // Lead to  "[42702] column reference "last_event_timestamp" is ambiguous"
     @Modifying
     @Query("""
         INSERT INTO job_record
@@ -50,7 +52,8 @@ public interface JobRecordRepository extends R2dbcRepository<JobRecord, String> 
         last_event_timestamp=:lastEventTimestamp, last_record_update_timestamp=:lastRecordUpdateTimestamp, status=:status, paused_bucket_key=:pausedBucketKey
         WHERE last_event_timestamp < :lastEventTimestamp AND id = :id
         """)
-    Mono<Integer> insertOnConflictUpdateAndCheckTime(
+    Mono<Long> insertOnConflictUpdateAndCheckTime(
         long id, Instant jobAcceptedTimestamp, Instant lastEventTimestamp,
         Instant lastRecordUpdateTimestamp, String status, String pausedBucketKey, long processId);
+    */
 }
